feat: answer type 정의
